Separate ScheduledEmail dialogs from DashboardHeader component
