+++ libs/sdk-ui-dashboard/src/presentation/dashboard/DashboardHeader/DashboardHeader.tsx
@@ -21,15 +21,12 @@ import {
     selectDashboardRef,
     selectDashboardShareInfo,
     selectDashboardTitle,
-    selectEnableInsightExportScheduling,
     selectEnableKPIDashboardExportPDF,
     selectEnableKPIDashboardSchedule,
     selectFilterContextFilters,
     selectIsLayoutEmpty,
     selectIsReadOnly,
     selectIsSaveAsDialogOpen,
-    selectIsScheduleEmailDialogOpen,
-    selectIsScheduleEmailManagementDialogOpen,
     selectMenuButtonItemsVisibility,
     selectPersistedDashboard,
     uiActions,
@@ -42,10 +39,13 @@ import { ExportDialogProvider } from "../../dialogs";
 
 import { downloadFile } from "../../../_staging/fileUtils/downloadFile";
 import { DefaultButtonBar, DefaultMenuButton, DefaultTopBar, IMenuButtonItem, TopBar } from "../../topBar";
-import { ScheduledEmailDialog, ScheduledEmailManagementDialog } from "../../scheduledEmail";
 import { SaveAsDialog } from "../../saveAs";
 import { DefaultFilterBar, FilterBar } from "../../filterBar";
 import { ShareDialogDashboardHeader } from "./ShareDialogDashboardHeader";
+import {
+    ScheduledEmailDialogProvider,
+    useScheduledEmailDialogProvider,
+} from "./ScheduledEmailDialogProvider";
 
 const useFilterBar = (): {
     filters: FilterContextItem[];
@@ -116,18 +116,9 @@ export const DashboardHeader = (): JSX.Element => {
     const { filters, onAttributeFilterChanged, onDateFilterChanged } = useFilterBar();
     const { title, onTitleChanged, onShareButtonClick, shareInfo } = useTopBar();
     const { addSuccess, addError, addProgress, removeMessage } = useToastMessage();
+    const { defaultOnScheduleEmailing } = useScheduledEmailDialogProvider();
 
     const dispatch = useDashboardDispatch();
-    const isScheduleEmailingDialogOpen = useDashboardSelector(selectIsScheduleEmailDialogOpen);
-    const isScheduleEmailingManagementDialogOpen = useDashboardSelector(
-        selectIsScheduleEmailManagementDialogOpen,
-    );
-    const openScheduleEmailingDialog = () => dispatch(uiActions.openScheduleEmailDialog());
-    const closeScheduleEmailingDialog = () => dispatch(uiActions.closeScheduleEmailDialog());
-    const openScheduleEmailingManagementDialog = () =>
-        dispatch(uiActions.openScheduleEmailManagementDialog());
-    const closeScheduleEmailingManagementDialog = () =>
-        dispatch(uiActions.closeScheduleEmailManagementDialog());
     const isSaveAsDialogOpen = useDashboardSelector(selectIsSaveAsDialogOpen);
     const openSaveAsDialog = () => dispatch(uiActions.openSaveAsDialog());
     const closeSaveAsDialog = () => dispatch(uiActions.closeSaveAsDialog());
@@ -189,26 +180,7 @@ export const DashboardHeader = (): JSX.Element => {
 
     const canCreateScheduledMail = useDashboardSelector(selectCanCreateScheduledMail);
     const isScheduledEmailingEnabled = !!useDashboardSelector(selectEnableKPIDashboardSchedule);
-    const isInsightExportSchedulingEnabled = useDashboardSelector(selectEnableInsightExportScheduling);
     const menuButtonItemsVisibility = useDashboardSelector(selectMenuButtonItemsVisibility);
-
-    /*
-     * exports and scheduling are not available when rendering a dashboard that is not persisted.
-     * this can happen when a new dashboard is created and is being edited.
-     *
-     * the setup of menu items available in the menu needs to reflect this.
-     */
-    const defaultOnScheduleEmailing = useCallback(() => {
-        if (!dashboardRef) {
-            return;
-        }
-
-        if (isInsightExportSchedulingEnabled) {
-            openScheduleEmailingManagementDialog();
-        } else {
-            openScheduleEmailingDialog();
-        }
-    }, [dashboardRef, isInsightExportSchedulingEnabled]);
 
     const defaultMenuItems = useMemo<IMenuButtonItem[]>(() => {
         if (!dashboardRef) {
@@ -270,34 +242,6 @@ export const DashboardHeader = (): JSX.Element => {
         isScheduledEmailingEnabled,
     ]);
 
-    const onScheduleEmailingError = useCallback(() => {
-        closeScheduleEmailingDialog();
-        addError({ id: "dialogs.schedule.email.submit.error" });
-    }, []);
-
-    const onScheduleEmailingSuccess = useCallback(() => {
-        closeScheduleEmailingDialog();
-        addSuccess({ id: "dialogs.schedule.email.submit.success" });
-    }, []);
-
-    const onScheduleEmailingCancel = useCallback(() => {
-        closeScheduleEmailingDialog();
-    }, []);
-
-    const onScheduleEmailingManagementAdd = useCallback(() => {
-        closeScheduleEmailingManagementDialog();
-        openScheduleEmailingDialog();
-    }, []);
-
-    const onScheduleEmailingManagementCancel = useCallback(() => {
-        closeScheduleEmailingManagementDialog();
-    }, []);
-
-    const onScheduleEmailingManagementError = useCallback(() => {
-        closeScheduleEmailingManagementDialog();
-        addError({ id: "dialogs.schedule.management.load.error" });
-    }, []);
-
     const onSaveAsError = useCallback(() => {
         closeSaveAsDialog();
         addError({ id: "messages.dashboardSaveFailed" });
@@ -316,22 +260,7 @@ export const DashboardHeader = (): JSX.Element => {
         <>
             <ToastMessages />
             <ExportDialogProvider />
-            {isScheduleEmailingManagementDialogOpen && (
-                <ScheduledEmailManagementDialog
-                    isVisible={isScheduleEmailingManagementDialogOpen}
-                    onAdd={onScheduleEmailingManagementAdd}
-                    onCancel={onScheduleEmailingManagementCancel}
-                    onError={onScheduleEmailingManagementError}
-                />
-            )}
-            {isScheduleEmailingDialogOpen && (
-                <ScheduledEmailDialog
-                    isVisible={isScheduleEmailingDialogOpen}
-                    onCancel={onScheduleEmailingCancel}
-                    onError={onScheduleEmailingError}
-                    onSuccess={onScheduleEmailingSuccess}
-                />
-            )}
+            <ScheduledEmailDialogProvider />
             <ShareDialogDashboardHeader />
             {isSaveAsDialogOpen && (
                 <SaveAsDialog
